refactor: AI型定義ファイルを分割してコード構成を改善
@app\src\domain\models\ai\types.tsを個別ファイルに分割：
- AISettings.ts: AI設定関連の型定義
- GameState.ts: ゲーム状態関連の型定義
- MoveTypes.ts: 手と評価関連の型定義
- PerformanceTypes.ts: パフォーマンス分析関連の型定義
- index.ts: 統一エクスポートポイント

影響を受けたファイル:
- PerformanceMetrics.ts: import文を'./index'に更新
- PerformanceMetrics.test.ts: import文を'./index'に更新
