+++ app/src/App.tsx
@@ -6,7 +6,7 @@ import type { InputPort } from './application/ports/InputPort'
 import type { PerformanceAnalysisService } from './application/services/PerformanceAnalysisService'
 import type { StrategyService } from './application/services/ai/StrategyService'
 import type { GameViewModel } from './application/viewmodels/GameViewModel'
-import type { AIMove, AISettings } from './domain/models/ai/types.ts'
+import type { AIMove, AISettings } from './domain/models/ai/index'
 import { defaultContainer } from './infrastructure/di/DefaultContainer'
 import { AIControlPanel } from './presentation/components/AIControlPanel'
 import { AIInsights } from './presentation/components/AIInsights'
