+++ app/src/application/services/ai/AIService.test.ts
@@ -4,10 +4,7 @@
 import { beforeEach, describe, expect, it } from 'vitest'
 
 import type { PuyoColor } from '../../../domain/models/Puyo'
-import type {
-  AIGameState,
-  AISettings,
-} from '../../../domain/models/ai/types.ts'
+import type { AIGameState, AISettings } from '../../../domain/models/ai/index'
 import { AIService } from './AIService'
 
 describe('AIService', () => {
