+++ app/src/application/services/visualization/ChartDataService.test.ts
@@ -3,7 +3,7 @@
  */
 import { describe, expect, test } from 'vitest'
 
-import type { PerformanceReport } from '../../../domain/models/ai/types'
+import type { PerformanceReport } from '../../../domain/models/ai/index'
 import type { PerformanceStatistics } from '../PerformanceAnalysisService'
 import { ChartDataService } from './ChartDataService'
