+++ app/src/infrastructure/adapters/PerformanceAdapter.ts
@@ -7,7 +7,7 @@ import type {
   GameSession,
   PerformanceData,
   PerformanceReport,
-} from '../../domain/models/ai/types'
+} from '../../domain/models/ai/index'
 
 /**
  * パフォーマンス分析インフラストラクチャアダプター
