+++ app/src/presentation/components/AIInsights.tsx
@@ -1,7 +1,7 @@
 /**
  * AI判断詳細表示コンポーネント
  */
-import type { AIMove, MoveEvaluation } from '../../domain/models/ai/types'
+import type { AIMove, MoveEvaluation } from '../../domain/models/ai/index'
 
 interface AIInsightsProps {
   /** 最後のAI判断 */
